rename auth service and user service
